requirements.txt
# Core essentials
fastapi==0.115.5
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Pipecat with required features (will install correct torch version)
pipecat-ai[silero,ultravox]==0.0.71

# Additional dependencies
librosa==0.10.1
requests>=2.31.0
websockets>=12.0
numpy>=1.26.4
scipy>=1.11.4

# For GPU support with Piper models (optional but recommended)
# onnxruntime-gpu>=1.16.3


